requirements.txt
numpy==1.24.4
pandas==2.0.3
prettytable==3.11.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
tzdata==2025.2
wcwidth==0.2.13
